* Added deathLostPercent for compatibility reasons (configurable + login.lua modification) * Changed doChangeGameState to doSetGameState by own reasons :P * Updated monster/Machines to 'energy' race * Fixed animatedtext.lua, shutdown.lua and yet updated serverinfo.lua * Updated function.lua * Declared IOLoginData object instead of retrieving it instance for every player when saving * Changed house_storage.data field type from BLOB (2G) to LONGBLOB (4G), to avoid any problems in future * Updated forgottenserver.s3db, cleanups * Patch Level 1
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@1203 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ item.cpp
@@ -861,7 +861,7 @@ std::string Item::getDescription(const ItemType& it, int32_t lookDistance, const
 				s << "Atk:";
 				if(it.abilities.elementType != COMBAT_NONE && it.decayTo < 1)
 				{
-					s << std::max(0, int32_t((item ? item->getAttack() : it.attack) - it.abilities.elementDamage));
+					s << std::max((int32_t)0, int32_t((item ? item->getAttack() : it.attack) - it.abilities.elementDamage));
 					if(it.extraAttack != 0 || (item && item->getExtraAttack() != 0))
 						s << " " << std::showpos << int32_t(item ? item->getExtraAttack() : it.extraAttack) << std::noshowpos;
 
@@ -1064,7 +1064,7 @@ std::string Item::getDescription(const ItemType& it, int32_t lookDistance, const
 	if(it.abilities.elementType != COMBAT_NONE && it.decayTo > 0)
 	{
 		s << std::endl << "It is temporarily enchanted with " << getCombatName(it.abilities.elementType) << " (";
-		s << std::max(0, int32_t((item ? item->getAttack() : it.attack) - it.abilities.elementDamage));
+		s << std::max((int32_t)0, int32_t((item ? item->getAttack() : it.attack) - it.abilities.elementDamage));
 		if(it.extraAttack != 0 || (item && item->getExtraAttack() != 0))
 			s << " " << std::showpos << int32_t(item ? item->getExtraAttack() : it.extraAttack) << std::noshowpos;
